requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
pyttsx3==2.90
gTTS==2.4.0
python-socketio==5.8.0
eventlet==0.33.3
Pillow==10.0.1
scikit-learn==1.3.0
